Move API-building code into dedicated dir
