+++ src/languages/rust/api/symbol_collection.rs
