+++ src/languages/rust/mod.rs
@@ -1,11 +1,6 @@
 mod analyser;
 mod api;
 mod metadata;
-mod namespace_construction;
-mod parsing;
-mod symbol_collection;
-mod symbol_resolution;
 mod test_helpers;
-mod types;
 
 pub use analyser::RustAnalyser;
